+++ ipax/__init__.py
@@ -73,4 +73,4 @@ __all__ = [
     "solve",
 ]
 
-__version__ = "0.6.0"
+__version__ = "0.6.1"